Refactor `types` iterators
